Initial design for supplemental page table entries
Without implementing further, this is what I think the supplemental page
table should contain: either an entry references a file (plus offset and
length), or a zeroed page. It also stores whether the page should be
writable.

When loading an executable and memory-mapping files, the SPT is filled
without actually loading the frames. When a page fault for an address
occurs, it is loaded on-demand using the information provided by the
SPT.