Add test which closes an mmapped file before writeback
This is in fact a very serious error. In the current implementation, if
the fd backing the mapping is closed before the changed mapping is
written back, the kernel accesses a freed struct file for the writeback.

If another file is opened immediately after the mapped one is closed, it
ends up at the same address, thus causing the writeback code to write to
the new file on unmmap.