Solve the "close mapped file before unmap" error
The file descriptors for mmapped files are no longer stored in the
process and can thus be no longer closed. Instead they are saved in
struct mape and closed once the mape is removed.

To prevent closing the executable's fd mapping which don't have to be
unmapped (i.e. everything from the executable) can be mapped via the
spt_map_segment() function which doesn't create a mape at all.