Change order of map_* functions in vm
The functions are now sorted in the order in that they are used.

First the segments of the executable have to be mapped with the help of
spt_map_segment(). If this fails there is no way to recover but to
terminate the process.

Then files may be mapped into memory by using spt_map_file(). This may
fail and full recovery is done automatically.

Lastly the mapped files (but not the segments) can be unmapped with
spt_unmap_file().

I hope this makes sense.